api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.4.0"
description = ""
authors = ["Anthony Miyaguchi <acmiyaguchi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
fastapi = "^0.78.0"
uvicorn = { extras = ["standard"], version = "^0.17.6" }
httpx = "^0.23.0"
simple-fast-python = "^1.0"
librosa = "^0.9.1"
matplotlib = "^3.5.2"
numpy = "^1.22.4"
PySoundFile = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
